Added first testFX test
