+++ pom.xml
@@ -5,17 +5,17 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>org.seniorglez</groupId>
-  <artifactId>karumiCodeTest</artifactId>
+  <artifactId>org.seniorglez.karumiCodeTest</artifactId>
   <version>1.0-SNAPSHOT</version>
 
-  <name>karumiCodeTest</name>
+  <name>org.seniorglez.karumiCodeTest</name>
   <url>https://www.karumi.com/</url>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.release>11</maven.compiler.release>
     <javafx.version>14.0.1</javafx.version>
-    <mainClassName>karumiCodeTest.App</mainClassName>
+    <mainClassName>org.seniorglez.karumiCodeTest.App</mainClassName>
   </properties>
 
   <dependencies>
@@ -45,6 +45,18 @@
       <groupId>org.testfx</groupId>
       <artifactId>testfx-junit5</artifactId>
       <version>4.0.16-alpha</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.testfx</groupId>
+      <artifactId>testfx-core</artifactId>
+      <version>4.0.16-alpha</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.hamcrest</groupId>
+      <artifactId>hamcrest</artifactId>
+      <version>2.1</version>
       <scope>test</scope>
     </dependency>
     <!--mocking libraries-->
