--- src/main/java/karumiCodeTest/views/login/LoginViewModel.java
@@ -1,4 +1,0 @@
-package karumiCodeTest.views.login;
-
-public class LoginViewModel {
-}
