+++ src/main/java/org/seniorglez/karumiCodeTest/App.java
@@ -1,10 +1,9 @@
-package karumiCodeTest;
+package org.seniorglez.karumiCodeTest;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 import java.io.IOException;
