+++ src/main/java/org/seniorglez/karumiCodeTest/services/login/LoginService.java
@@ -16,27 +16,37 @@
  *  TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  *  IN THE SOFTWARE.
  */
-package karumiCodeTest;
+package org.seniorglez.karumiCodeTest.services.login;
 
-import karumiCodeTest.config.PropertiesReader;
-import org.junit.jupiter.api.Test;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import javafx.concurrent.Service;
+import javafx.concurrent.Task;
+import org.seniorglez.karumiCodeTest.model.Credentials;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-public class TestPropertiesReader {
+/**
+ * A service which wraps the {@link LoginTask}.
+ *
+ * Before start the {@link LoginService} make sure that there is a {@link Credentials} object stored on the instance.
+ */
+public class LoginService extends Service<String> {
 
     /**
-     *
-     * https://regexr.com/37i6s
+     * A object which represents the credentials used to log in the server.
      */
-    @Test
-    public void shouldRetrieveAnURL() {
-        PropertiesReader propertiesReader = PropertiesReader.instanciate();
-        String url = propertiesReader.getProerty("url");
-        Pattern pattern = Pattern.compile("https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{2,256}\\.[a-z]{2,4}\\b([-a-zA-Z0-9@:%_\\+.~#?&//=]*)");
-        Matcher matcher = pattern.matcher(url);
-        assertTrue(matcher.matches());
+    private Credentials credentials;
+
+    /**
+     * Invoked after the service is started on the JavaFX Application Thread.
+     * @return A {@link LoginTask} constructed with the {@link Credentials} object stored on the instance.
+     */
+    @Override
+    protected Task<String> createTask() {
+        return new LoginTask(credentials);
+    }
+
+    /** Set the {@link Credentials} object which is used to intanciate the LoginTask when the {@link LoginService} starts it's execution.
+     * @param credentials the new credentials.
+     */
+    public void setCredentials(Credentials credentials) {
+        this.credentials = credentials;
     }
 }
