+++ src/main/java/org/seniorglez/karumiCodeTest/views/login/LoginViewModel.java
@@ -1,0 +1,4 @@
+package org.seniorglez.karumiCodeTest.views.login;
+
+public class LoginViewModel {
+}
