+++ src/main/resources/org/seniorglez/karumiCodeTest/config.properties
