+++ src/main/resources/org/seniorglez/karumiCodeTest/views/login/logo.png
